feat(manager-components): split translation into two namespaces
ref: MANAGER-15076
