+++ packages/manager-components/src/components/region/translations/region/index.ts
